+++ crates/server/raft-log/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "databend-meta-raft-log"
+description = "WAL-backed raft log"
+version = { workspace = true }
+authors = { workspace = true }
+license = { workspace = true }
+publish = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+anyerror = { workspace = true }
+databend-meta-types = { workspace = true }
+deepsize = { workspace = true }
+display-more = { workspace = true }
+log = { workspace = true }
+raft-log = { workspace = true }
+rmp-serde = { workspace = true }
+serde = { workspace = true }
+thiserror = { workspace = true }
+tokio = { workspace = true }
+
+[dev-dependencies]
+anyhow = { workspace = true }
+pretty_assertions = { workspace = true }
+tempfile = { workspace = true }
+
+[lints]
+workspace = true
