+++ crates/server/raft-log/src/codec_wrapper.rs
