+++ crates/server/service/Cargo.toml
@@ -20,6 +20,7 @@ databend-meta-client = { workspace = true }
 databend-meta-kvapi = { workspace = true }
 databend-meta-leveled-store = { workspace = true }
 databend-meta-raft-config = { workspace = true }
+databend-meta-raft-log = { workspace = true }
 databend-meta-raft-store = { workspace = true }
 databend-meta-runtime-api = { workspace = true }
 databend-meta-sled-store = { workspace = true }
