+++ crates/server/service/src/meta_node/meta_node.rs
@@ -28,7 +28,7 @@ use databend_meta_raft_config::MetaStartupError;
 use databend_meta_raft_config::StateMachineFeature;
 use databend_meta_raft_config::config::RaftConfig;
 use databend_meta_raft_config::data_version::DATA_VERSION;
-use databend_meta_raft_store::log_store::RaftLogStat;
+use databend_meta_raft_log::RaftLogStat;
 use databend_meta_raft_store::utils::seq_marked_to_seqv;
 use databend_meta_runtime_api::JoinHandle;
 use databend_meta_runtime_api::SpawnApi;
