+++ crates/server/service/src/store/store.rs
@@ -22,12 +22,12 @@ use std::time::Duration;
 use anyerror::AnyError;
 use databend_meta_raft_config::MetaStartupError;
 use databend_meta_raft_config::config::RaftConfig;
+use databend_meta_raft_log::Cw;
+use databend_meta_raft_log::RaftLog;
+use databend_meta_raft_log::log_store_meta::LogStoreMeta;
+use databend_meta_raft_log::util;
 use databend_meta_raft_store::db_exporter::DBExporter;
 use databend_meta_raft_store::header::Header;
-use databend_meta_raft_store::log_store;
-use databend_meta_raft_store::log_store::Cw;
-use databend_meta_raft_store::log_store::RaftLog;
-use databend_meta_raft_store::log_store::util;
 use databend_meta_raft_store::ondisk::TREE_HEADER;
 use databend_meta_raft_store::sled_compat::key_spaces::RaftStoreEntry;
 use databend_meta_raft_store::snapshot_store::MetaSnapshotId;
@@ -111,7 +111,7 @@ impl<SP: SpawnApi> RaftStore<SP> {
         let id = stored_node_id.unwrap_or(config.id);
 
         if !is_opened {
-            log.save_user_data(Some(log_store::log_store_meta::LogStoreMeta {
+            log.save_user_data(Some(LogStoreMeta {
                 node_id: Some(config.id),
             }))
             .map_err(to_startup_err)?;
