+++ crates/server/service/tests/it/store.rs
@@ -18,8 +18,8 @@ use std::time::Duration;
 use databend_meta::meta_node::meta_node::LogStore;
 use databend_meta::meta_node::meta_node::SMStore;
 use databend_meta::store::RaftStore;
+use databend_meta_raft_log::util::blocking_flush;
 use databend_meta_raft_store::db_exporter::DBExporter;
-use databend_meta_raft_store::log_store::util::blocking_flush;
 use databend_meta_raft_store::snapshot_store::MetaSnapshotId;
 use databend_meta_runtime_api::TokioRuntime;
 use databend_meta_snapshot_db::DB;
